GIT提交訊息：
feat: 重構身份驗證並整合 js-cookie

- 新增 `js-cookie` 套件以處理cookies
- 新增 `@types/js-cookie` 型別定義以支援TypeScript
- 用TypeScript建立一個新的`LoginForm`組件，具有表單驗證和樣式
- 將登入表單從`page.tsx`提取至`LoginForm.tsx`
- 為小型設備調整在`signup/page.tsx`中圖片顯示條件
- 更新API方法以利用`js-cookie`處理取得和設定`token`
