+++ package.json
@@ -20,6 +20,7 @@
     "@mui/material": "^5.15.3",
     "@mui/material-nextjs": "^5.15.3",
     "@uidotdev/usehooks": "^2.4.1",
+    "js-cookie": "^3.0.5",
     "next": "^14.1.0",
     "react": "^18",
     "react-dom": "^18",
@@ -29,6 +30,7 @@
   },
   "devDependencies": {
     "@svgr/webpack": "^8.1.0",
+    "@types/js-cookie": "^3",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
