+++ src/assets/api/index.ts
@@ -1,7 +1,8 @@
+import Cookies from 'js-cookie';
+
 const baseUrl = process.env.NEXT_PUBLIC_API_URL;
 
-const token =
-  'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NWFlMDRkYzE0ZjM5NmUwNTlhOWM3M2EiLCJpYXQiOjE3MDYyMDE1ODksImV4cCI6MTcwNjgwNjM4OX0.34299tu2787_3J0RG7lJ4dZpiCQNyNcOPkuNXqnbUAk';
+const token = Cookies.get('token') || '';
 
 export async function getUser() {
   const res = await fetch(`${baseUrl}/api/v1/user`, {
@@ -15,6 +16,7 @@ export async function getUser() {
   if (!res.ok) {
     throw new Error('Failed to fetch data');
   }
+  res.json().then((data) => data.token && Cookies.set('token', data.token));
 
   return res.json();
 }
@@ -78,6 +80,7 @@ export async function userLogin(data: UserLoginData) {
   if (!res.ok) {
     throw new Error('Failed to fetch data');
   }
+  res.json().then((data) => data.token && Cookies.set('token', data.token));
   return res.json();
 }
 
@@ -92,6 +95,7 @@ export async function userRegister(data: UserRegisterData) {
   if (!res.ok) {
     throw new Error('Failed to fetch data');
   }
+  res.json().then((data) => data.token && Cookies.set('token', data.token));
   return res.json();
 }
